requirements.txt
streamlit
Pillow
numpy
opencv-python
mediapipe